wip hats

+++ froggyfocus/Character/FrogAppearanceData.cs
@@ -1,6 +1,6 @@
 public class FrogAppearanceData
 {
-    public int ColorR { get; set; }
-    public int ColorG { get; set; }
-    public int ColorB { get; set; }
+    public float ColorR { get; set; }
+    public float ColorG { get; set; }
+    public float ColorB { get; set; }
 }
